refactor: context 상태관리 App.jsx로 이동 사유: 상태를 전역적으로 사용 불가했기 때문
